Fix #1975: Split graph lib.rs (4,278 lines) into focused submodules
Extract the monolithic GraphStore impl block and its 103 tests into
four submodules following the same pattern used in the vector crate's
store.rs split (#1961):

- lifecycle.rs: create/delete/list/meta graph operations (672 lines)
- nodes.rs: node CRUD, ref/type index queries (1,045 lines)
- edges.rs: edge CRUD, neighbors, traversal helpers, counters (882 lines)
- bulk.rs: bulk_insert, batch_add_edges (1,644 lines)

lib.rs reduced from 4,278 to 108 lines (struct + constructor +
snapshot/stats/adjacency). No public API changes — pure structural
refactoring. Two private helpers (read/write_edge_type_count) widened
to pub(crate) for cross-module access.
